// src/generated/gravatar-api/models/Interest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gravatar Public API
 * Gravatar\'s public API endpoints
 *
 * The version of the OpenAPI document: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * An interest the user has added to their profile.
 * @export
 * @interface Interest
 */
export interface Interest {
    /**
     * The unique identifier for the interest.
     * @type {number}
     * @memberof Interest
     */
    id: number;
    /**
     * The name of the interest as originally defined (most often in English).
     * @type {string}
     * @memberof Interest
     */
    name: string;
}

/**
 * Check if a given object implements the Interest interface.
 */
export function instanceOfInterest(value: object): value is Interest {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function InterestFromJSON(json: any): Interest {
    return InterestFromJSONTyped(json, false);
}

export function InterestFromJSONTyped(json: any, ignoreDiscriminator: boolean): Interest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
    };
}

export function InterestToJSON(json: any): Interest {
    return InterestToJSONTyped(json, false);
}

export function InterestToJSONTyped(value?: Interest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
    };
}